Create Authentication Hooks
