+++ src/components/NewRoom.tsx
@@ -4,12 +4,12 @@ import '../styles/auth.scss';
 import { Button } from '../components/Button';
 import illustrationImg from '../assets/images/illustration.svg';
 import logoImg from '../assets/images/logo.svg';
-import { AuthContext } from '../App';
+import { useAuth } from '../hooks/useAuth';
 
 
 export function NewRoom() {
 
-  const { user } = useContext(AuthContext);
+  const { user } = useAuth();
 
   return (
     <div id="page-auth">
@@ -21,7 +21,6 @@ export function NewRoom() {
       <main>
         <div className="main-content">
           <img src={logoImg} alt="letmeask logo" />
-          <h1>{user?.name}</h1>
           <h2>Create new Room</h2>
           <div></div>
           <form action="">
