+++ src/hooks/useAuth.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import {AuthContext } from '../contexts/AuthContext';
+
+export function useAuth() {
+  const value = useContext(AuthContext)
+
+  return value;
+}
