+++ src/pages/Home.tsx
@@ -1,7 +1,5 @@
 import { useHistory } from 'react-router-dom' //create path for buttton
-import { useContext } from 'react'; //context I wanna use from CreateContext in App
-
-import { AuthContext } from '../App'
+import { useAuth } from '../hooks/useAuth';
 
 import illustrationImg from '../assets/images/illustration.svg';
 import logoImg from '../assets/images/logo.svg';
@@ -12,7 +10,7 @@ import { Button } from '../components/Button';
 export function Home() {
 
   const history = useHistory(); //HOOK - all hooks must be inside the component - navigation
-  const { user, signInWithGoogle } = useContext(AuthContext);
+  const { user, signInWithGoogle } = useAuth();
 
   async function handleCreateRoom() {
     if (!user) {
